requirements.txt
python-dotenv==1.0.0
flask==2.2.5
gunicorn==20.1.0
pymongo==4.4.0
kafka-python==2.0.2